+++ TakeoutExtractor.Gui/MainPage.xaml.cs
@@ -14,18 +14,12 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
         public MainPage()
         {
             InitializeComponent();
-        }
 
-
-
-        protected override void OnAppearing()
-        {
-#if DEBUG
-            InputDirEntry.Text = "D:\\Temp\\takeout-20220331T163017Z-001\\Takeout";
-            OutputDirEntry.Text = "D:\\Temp\\TakeoutData";
-#endif
+            // Global controls
             CreateLogFileCbx.IsChecked = GlobalOptions.Defaults.CreateLogFile;
+            StopOnErrorCbx.IsChecked = GlobalOptions.Defaults.StopOnError;
 
+            // Phot controls.
             PhotosExtractCbx.IsChecked = true;
             PhotosFileNameFormatTxt.Text = !string.IsNullOrEmpty(PhotoOptions.Defaults.OutputFileNameFormat) ? PhotoOptions.Defaults.OutputFileNameFormat : "";
             PhotosUpdateExifCbx.IsChecked = PhotoOptions.Defaults.UpdateExif;
@@ -33,6 +27,8 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
             PhotosSuffixOriginalsTxt.Text = !string.IsNullOrEmpty(PhotoOptions.Defaults.OriginalsSuffix) ? PhotoOptions.Defaults.OriginalsSuffix : "";
             PhotosSubdirOriginalsTxt.Text = !string.IsNullOrEmpty(PhotoOptions.Defaults.OriginalsSubdirName) ? PhotoOptions.Defaults.OriginalsSubdirName : "";
             PhotosSubdirOrganisationPicker.SelectedIndex = (int)PhotoOptions.Defaults.OrganiseBy;
+            // Do a chage event on the keep originals checkbox to ensure that associated controls are correctly enabled/disabled.
+            OnPhotosKeepOriginalsChanged(this, new CheckedChangedEventArgs(PhotosKeepOriginalsCbx.IsChecked));
         }
 
 
@@ -45,7 +41,7 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
         {
             var msg = string.Format("Takeout Extractor v{0} by Andy Johnson. See https://github.com/andyjohnson0/TakeoutExtractor for info.",
                                     Assembly.GetExecutingAssembly().GetName().Version!.ToString());
-            DisplayAlert("About TakeoutExtractor", msg, "Ok");
+            DisplayAlert("About Takeout Extractor", msg, "Ok");
         }
 
 
@@ -90,21 +86,16 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
             EventArgs e)
         {
             // Show the overlay that give feedback progress.
-            MainGrid.IsEnabled = false;
-            ProgressOverlay? progressOverlay = new ProgressOverlay();
-            progressOverlay.ZIndex = 99;
-            // TODO: Improve how this is added to the layout.
-            progressOverlay.SetValue(Grid.RowProperty, 0);
-            progressOverlay.SetValue(Grid.ColumnProperty, 0);
-            MainGrid.Children.Add(progressOverlay);
-            //progressOverlay.IsEnabled = true;
+            var progressOverlay = new ProgressOverlay();
+            progressOverlay.Show(MainGrid);
 
             // Set-up extractor and options.
             var globalOptions = new GlobalOptions()
             {
                 InputDir = new DirectoryInfo(InputDirEntry.Text),
                 OutputDir = new DirectoryInfo(OutputDirEntry.Text),
-                CreateLogFile = CreateLogFileCbx.IsChecked
+                CreateLogFile = CreateLogFileCbx.IsChecked,
+                StopOnError = StopOnErrorCbx.IsChecked
             };
             var mediaOptions = new List<IExtractorOptions>();
             if (this.PhotosExtractCbx.IsChecked)
@@ -141,7 +132,8 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
             try
             {
                 var results = await Task.Run(() => extractor.ExtractAsync(cancellationTokenSource.Token));
-                await DisplayAlert("Completed", "The extraction operation completed successfully" + Environment.NewLine + Environment.NewLine + FormatResults(results), "Ok");
+                progressOverlay.Close();
+                await DisplayResults(results);
             }
             catch(OperationCanceledException)
             {
@@ -154,19 +146,29 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
             finally
             {
                 // Hide the overlay.
-                progressOverlay.IsVisible = false;
+                progressOverlay.Close();
                 progressOverlay = null;
-                MainGrid.IsEnabled = true;
             }
         }
 
 
-        private static string FormatResults(
+        private async Task DisplayResults(
             IEnumerable<IExtractorResults>? results)
         {
             var sb = new StringBuilder();
 
-            sb.Append("Results:" + Environment.NewLine);
+            sb.AppendLine("The extraction operation completed");
+            sb.AppendLine();
+
+            sb.AppendLine("Results:");
+            var alerts = results?.SelectMany(a => a.Alerts);
+            if (alerts != null)
+            {
+                var errorCount = alerts.Count(a => a.Type == ExtractorAlertType.Error);
+                var warningCount = alerts.Count(a => a.Type == ExtractorAlertType.Warning);
+                var infoCount = alerts.Count(a => a.Type == ExtractorAlertType.Information);
+                sb.AppendLine($"{errorCount} error, {warningCount} warning, {infoCount} information");
+            }
             if (results != null && results.Any())
             {
                 foreach (var result in results)
@@ -175,17 +177,22 @@ namespace uk.andyjohnson.TakeoutExtractor.Gui
                         sb.Append("Photos and Videos: ");
                     else
                         sb.Append("Other: ");
-                    sb.Append(string.Format("{0}% in {1}",
+                    sb.AppendLine(string.Format("{0}% in {1}",
                         result.Coverage * 100M,
                         result.Duration.ToString(@"hh\h\ mm\m\ ss\s")));
                 }
             }
             else
             {
-                sb.Append("None");
+                sb.AppendLine("None");
             }
 
-            return sb.ToString();
+            // Display results. If we have alerts then add a details button to display them.
+            var choice = await QuestionDialog.ShowAsync("Completed", sb.ToString(), "Ok", alerts?.Count() > 0 ? "Details" : null);
+            if (choice != null && choice == "Details")
+            {
+                await Navigation.PushAsync(new AlertsPage(alerts!));
+            }
         }
     }
 }