+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 #Change Log
 
+##2022-08-25 v0.5
+
+Extractors return results and error/warning/info objects that are presented and logged. Option to stop on first error.
+
+Various fixed and code cleanup.
+
+
+##2022-08-12 v0.4
+
+Added option for extracting photos into directory hierarchy based on creation date (year, year/month, year/month/day)
+
+Added options to create a JSON log file describing what was extracted.
+
+General code clean-up
+
+
 ##2022-08-06 v0.3
 
 Initial, minimal installable GUI
+
 
 ##2022-07-27 v0.2